+++ scanner/src/main/java/ca/on/oicr/gsi/runscanner/scanner/processor/PathType.java
@@ -2,5 +2,6 @@ package ca.on.oicr.gsi.runscanner.scanner.processor;
 
 public enum PathType {
   DIRECTORY,
-  FILE
+  FILE,
+  VIRTUAL
 }